requirements.txt
opencv_python
Pillow
streamlit
ultralytics
